Add database module and fix existing errors
The `Database` struct and its associated logic was moved to its own
module to create structure and separation of concerns.

The application has new logging capabilities via functionality provided
by the `tracing` and `tracing-subscriber` crates. The logging module
enables the server to collect telemetry data, display it to the console
(stdout/stderr) and write it to a designated log file. This is very
useful for debugging and testing during development and a necessity for
production software.

Temporary changes and fixes were added to simplify existing code and to
get rid of lifetime bounds. This creates many unnecessary and/or
expensive allocations and will be addressed by modifying the storage
engine to work with `Bytes` rather than `String` data.